requirements_training.txt
numpy>=1.21.0
tensorflow>=2.12.0
matplotlib>=3.5.0
Pillow>=8.3.0